% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{data_add}
\alias{data_add}
\title{Adds a new resource to the KiProject. The resource can be a remote or local file or directory. If the resource already exists it will be updated with the provided parameters.}
\usage{
data_add(ki_project, remote_uri_or_local_path, name = NULL,
  version = NULL, data_type = NULL)
}
\arguments{
\item{ki_project}{a Ki Project obtained from \code{\link{ki_project}}}

\item{remote_uri_or_local_path}{The remote URI (e.g., syn:syn123456) or local path of the directory for file.}

\item{name}{A user friendly name for the resource.}

\item{version}{The version of the file to add.}

\item{data_type}{The DataType of the file. This is only required when a remote_uri is provided and the remote}
}
\value{
KiProjectResource
}
\description{
Adds a new resource to the KiProject. The resource can be a remote or local file or directory. If the resource already exists it will be updated with the provided parameters.
}
\note{
This can be used in either of the following ways:
\preformatted{
ki_project$data_add(...)
data_add(ki_project, ...)
}
}
