% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{data_change}
\alias{data_change}
\title{Changes the name or version on a KiProjectResource.}
\usage{
data_change(ki_project, resource_or_identifier, name = NULL,
  version = NULL)
}
\arguments{
\item{ki_project}{a Ki Project obtained from \code{\link{ki_project}}}

\item{resource_or_identifier}{KiProjectResource object or a valid identifier (local path, remote URI, name).}

\item{name}{The new name.}

\item{version}{The new version (or 'None' to clear the version).}
}
\value{
KiProjectResource
}
\description{
Changes the name or version on a KiProjectResource.
}
\note{
This can be used in either of the following ways:
\preformatted{
ki_project$data_change(...)
data_change(ki_project, ...)
}
}
