% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{data_list}
\alias{data_list}
\title{Prints out a table of all the resources in the KiProject.}
\usage{
data_list(ki_project, all = FALSE)
}
\arguments{
\item{ki_project}{a Ki Project obtained from \code{\link{ki_project}}}

\item{all}{Set to True to include all child resources.}
}
\value{
BeautifulTable
}
\description{
Prints out a table of all the resources in the KiProject.
}
\note{
This can be used in either of the following ways:
\preformatted{
ki_project$data_list(...)
data_list(ki_project, ...)
}
}
