% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-auto.R
\name{data_pull}
\alias{data_pull}
\title{Downloads a specific resource or all resources in the KiProject.}
\usage{
data_pull(ki_project, resource_or_identifier = NULL)
}
\arguments{
\item{ki_project}{a Ki Project obtained from \code{\link{ki_project}}}

\item{resource_or_identifier}{KiProjectResource object or a valid identifier (local path, remote URI, name).}
}
\value{
The absolute path to the pulled resource or a list of absolute paths for all pulled resources.
}
\description{
Downloads a specific resource or all resources in the KiProject.
}
\note{
This can be used in either of the following ways:
\preformatted{
ki_project$data_pull(...)
data_pull(ki_project, ...)
}
}
