% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{data_push}
\alias{data_push}
\title{Uploads a specific resource or all local non-pushed resources.}
\usage{
data_push(ki_project, resource_or_identifier = NULL)
}
\arguments{
\item{ki_project}{a Ki Project obtained from \code{\link{ki_project}}}

\item{resource_or_identifier}{KiProjectResource object or a valid identifier (local path, remote URI, name).}
}
\value{
The absolute path to the pushed resource or a list of absolute paths for all pushed resources.
}
\description{
Uploads a specific resource or all local non-pushed resources.
}
\note{
This can be used in either of the following ways:
\preformatted{
ki_project$data_push(...)
data_push(ki_project, ...)
}
}
