% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{data_remove}
\alias{data_remove}
\title{Removes a resource from the KiProject. This does not delete the file locally or remotely, it is only removed from the KiProject manifest.}
\usage{
data_remove(ki_project, resource_or_identifier)
}
\arguments{
\item{ki_project}{a Ki Project obtained from \code{\link{ki_project}}}

\item{resource_or_identifier}{KiProjectResource object or a valid identifier (local path, remote URI, name).}
}
\value{
KiProjectResource
}
\description{
Removes a resource from the KiProject. This does not delete the file locally or remotely, it is only removed from the KiProject manifest.
}
\note{
This can be used in either of the following ways:
\preformatted{
ki_project$data_remove(...)
data_remove(ki_project, ...)
}
}
