% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-auto.R
\name{data_type_from_project_path}
\alias{data_type_from_project_path}
\title{Gets the DataType from a local path. The local path must be in one of the KiProject's DataType directories.}
\usage{
data_type_from_project_path(ki_project, local_path)
}
\arguments{
\item{ki_project}{a Ki Project obtained from \code{\link{ki_project}}}

\item{local_path}{Path to get the DataType from.}
}
\value{
The DataType or NULL.
}
\description{
Gets the DataType from a local path. The local path must be in one of the KiProject's DataType directories.
}
\note{
This can be used in either of the following ways:
\preformatted{
ki_project$data_type_from_project_path(...)
data_type_from_project_path(ki_project, ...)
}
}
