% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{data_type_to_project_path}
\alias{data_type_to_project_path}
\title{Gets the absolute path to the DataType directory in the KiProject.}
\usage{
data_type_to_project_path(ki_project, data_type)
}
\arguments{
\item{ki_project}{a Ki Project obtained from \code{\link{ki_project}}}

\item{data_type}{The DataType to get the path for.}
}
\value{
Absolute path to the local directory as a string.
}
\description{
Gets the absolute path to the DataType directory in the KiProject.
}
\note{
This can be used in either of the following ways:
\preformatted{
ki_project$data_type_to_project_path(...)
data_type_to_project_path(ki_project, ...)
}
}
