% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-auto.R
\name{find_missing_resources}
\alias{find_missing_resources}
\title{Finds all local DataType directories and files that have not been added to the KiProject resources.}
\usage{
find_missing_resources(ki_project)
}
\arguments{
\item{ki_project}{a Ki Project obtained from \code{\link{ki_project}}}
}
\value{
List of paths
}
\description{
Finds all local DataType directories and files that have not been added to the KiProject resources.
}
\note{
This can be used in either of the following ways:
\preformatted{
ki_project$find_missing_resources(...)
find_missing_resources(ki_project, ...)
}
}
