% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{find_project_resources_by}
\alias{find_project_resources_by}
\title{Finds all resources in the KiProject by any of KiProjectResource attributes.}
\usage{
find_project_resources_by(ki_project, operator = "and", ...)
}
\arguments{
\item{ki_project}{a Ki Project obtained from \code{\link{ki_project}}}

\item{operator}{The operator to use when finding by more than one attribute. Must be one of: 'and', 'or'.}

\item{...}{KiProjectResource attributes and values to find by.}
}
\value{
List of KiProjectResources or an empty list.
}
\description{
Finds all resources in the KiProject by any of KiProjectResource attributes.
}
\note{
This can be used in either of the following ways:
\preformatted{
ki_project$find_project_resources_by(...)
find_project_resources_by(ki_project, ...)
}
}
