% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{is_configured}
\alias{is_configured}
\title{Check to see if the kitools Python environment has been configured}
\usage{
is_configured(msg = message)
}
\arguments{
\item{msg}{What function to use for messages (could be called at package startup or elsewhere in the package)}
}
\description{
Check to see if the kitools Python environment has been configured
}
