% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{is_project_data_type_path}
\alias{is_project_data_type_path}
\title{Gets if the local_path is in one of the DataType directories.}
\usage{
is_project_data_type_path(ki_project, local_path)
}
\arguments{
\item{ki_project}{a Ki Project obtained from \code{\link{ki_project}}}

\item{local_path}{Path to check.}
}
\value{
True or False
}
\description{
Gets if the local_path is in one of the DataType directories.
}
\note{
This can be used in either of the following ways:
\preformatted{
ki_project$is_project_data_type_path(...)
is_project_data_type_path(ki_project, ...)
}
}
