% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{kitools_configure}
\alias{kitools_configure}
\title{One-time configuration of environment for kitools}
\usage{
kitools_configure()
}
\description{
One-time configuration of environment for kitools
}
\details{
This installs an isolated Python distribution along with required dependencies so that the kitools R package can seamlessly wrap the kitools Python package.
}
