% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{kitools_update}
\alias{kitools_update}
\title{Update kitools environment configuration}
\usage{
kitools_update()
}
\description{
Update kitools environment configuration
}
\details{
This updates the kitools Python dependencies.
}
