% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-auto.R
\name{load}
\alias{load}
\title{Loads the KiProject from a config file.}
\usage{
load(ki_project)
}
\arguments{
\item{ki_project}{a Ki Project obtained from \code{\link{ki_project}}}
}
\value{
True if the config file exists and was loaded.
}
\description{
Loads the KiProject from a config file.
}
\note{
This can be used in either of the following ways:
\preformatted{
ki_project$load(...)
load(ki_project, ...)
}
}
