% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-auto.R
\name{save}
\alias{save}
\title{Saves the KiProject to a config file.}
\usage{
save(ki_project)
}
\arguments{
\item{ki_project}{a Ki Project obtained from \code{\link{ki_project}}}
}
\description{
Saves the KiProject to a config file.
}
\note{
This can be used in either of the following ways:
\preformatted{
ki_project$save(...)
save(ki_project, ...)
}
}
