% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{show_missing_resources}
\alias{show_missing_resources}
\title{Shows all local DataType directories and files that have not been added to the KiProject resources.}
\usage{
show_missing_resources(ki_project)
}
\arguments{
\item{ki_project}{a Ki Project obtained from \code{\link{ki_project}}}
}
\description{
Shows all local DataType directories and files that have not been added to the KiProject resources.
}
\note{
This can be used in either of the following ways:
\preformatted{
ki_project$show_missing_resources(...)
show_missing_resources(ki_project, ...)
}
}
