% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{syn_login}
\alias{syn_login}
\title{One-time setup of Synapse credentials}
\usage{
syn_login(email = NULL, pass = NULL)
}
\arguments{
\item{email}{your Synapse email address (if not supplied, will be prompted)}

\item{pass}{your Synapse password (if not supplied, will be prompted)}
}
\description{
One-time setup of Synapse credentials
}
