% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ki_project.R
\name{ki_project}
\alias{ki_project}
\title{Initialize a kitools project}
\usage{
ki_project(local_path, title = NULL, description = NULL,
  project_uri = NULL, resources = NULL)
}
\arguments{
\item{local_path}{local_path}

\item{title}{title}

\item{description}{description}

\item{project_uri}{project_uri}

\item{resources}{resources}
}
\description{
Initialize a kitools project
}
